tilelink2: move general-purpose code out of tilelink2 package
