+++ src/main/scala/rocketchip/Periphery.scala
@@ -6,6 +6,7 @@ import Chisel._
 import cde.{Parameters, Field}
 import junctions._
 import junctions.NastiConstants._
+import diplomacy._
 import uncore.tilelink._
 import uncore.tilelink2._
 import uncore.converters._
