+++ src/main/scala/uncore/tilelink2/Fragmenter.scala
@@ -4,6 +4,7 @@ package uncore.tilelink2
 
 import Chisel._
 import chisel3.internal.sourceinfo.SourceInfo
+import diplomacy._
 import scala.math.{min,max}
 
 // minSize: minimum size of transfers supported by all outward managers
