+++ src/main/scala/uncore/tilelink2/Isolation.scala
@@ -4,6 +4,7 @@ package uncore.tilelink2
 
 import Chisel._
 import chisel3.internal.sourceinfo.SourceInfo
+import diplomacy._
 
 class TLIsolation(f: (Bool, UInt) => UInt) extends LazyModule
 {
